requirements.txt
cudatoolkit==11.3.1
numpy==1.19.2
python==3.6.13
pytorch==1.10.1
tqdm==4.64.1
transformers==2.8.0
